src/main/java/com/anymind/pos/config/JacksonConfig.java
package com.anymind.pos.config;

import com.anymind.pos.config.serialization.PriceDeserializer;
import com.anymind.pos.config.serialization.PriceSerializer;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.module.SimpleModule;
import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;

import java.math.BigDecimal;

@Configuration
public class JacksonConfig {

    @Bean
    public ObjectMapper objectMapper() {
        ObjectMapper objectMapper = new ObjectMapper();
        SimpleModule module = new SimpleModule();

        module.addSerializer(BigDecimal.class, new PriceSerializer());
        module.addDeserializer(BigDecimal.class, new PriceDeserializer());

        objectMapper.registerModule(module);
        objectMapper.registerModule(new Jdk8Module());
        objectMapper.registerModule(new JavaTimeModule());
        return objectMapper;
    }
}
